+++ src/components/admin/EditProduct.tsx
@@ -213,7 +213,7 @@ const EditProduct = ({ categories, product, type, getProducts}: {categories: any
                         <Input
                             type="number"
                             value={createForm.packSize ?? ""}
-                            onChange={(e) => setCreateForm((f) => ({ ...f, packSize: parseInt(e.target.value) || 0 }))}
+                            onChange={(e) => setCreateForm((f) => ({ ...f, packSize: parseFloat(e.target.value) || 0 }))}
                             placeholder="3"
                             className="h-10 rounded-xl"
                         />
